+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed:
+- Text color for entry duration when timer running
+
+## [0.8.239] - 2023-01-14
 ### Changed:
 - Long press on entry type filter to select only one type
 - Select all entry types toggle button